requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
redis==5.0.1
python-dotenv==1.0.0
colorama==0.4.6
pytest==7.4.4